test(segmented-control): recategorize and clean up browser tests
